fix attempts

+++ models/danet_lib/lib/utils.py
@@ -192,6 +192,6 @@ def define_device(device_name):
 
 
 def normalize_reg_label(label, mu, std):
-    norm_label = ((label - mu) / std).astype(np.float32)
+    norm_label = ((label - mu) / std).astype(float)
     norm_label = norm_label.reshape(-1, 1)
     return norm_label
